safari-600.7-branch/LayoutTests/fast/viewport/viewport-legacy-ordering-4-expected.txt

ALERT: viewport size 300x330 scale 1.06667 with limits [1.06667, 5] and userScalable true
